Add a part to check for production log
graby should now log an error message when the request fail. When it does that, the whole logs related to the request will be written in the production log. It might help debugging the problem.

Also, I moved some part of the error documentation so user will have to check for error message in production log, then enable dev log and finally try to fix the problem by writing a siteconfig.