More actionable Docker launch command in installation docs
This improves the docker command to launch the container. The `docker pull` command only fetches the container, but does not launch it.

With the updated, a Docker user can quickly spin up a Wallabag instance for evaluation. For that reason, I chose port 8080 on the host machine since port 80 is often already taken. Additionally, wallabag data is persisted using Docker named volumes instead of mounting absolute paths on the host machine that may not exist.